requirements.txt
python-telegram-bot[webhooks]==20.8
stripe
openai